restructure pred in the works
