+++ frontend/src/Language/Granule/Checker/Types.hs
@@ -11,7 +11,7 @@ import Control.Monad.State.Strict
 import Language.Granule.Checker.Constraints.Compile
 
 import Language.Granule.Checker.Effects
-import Language.Granule.Checker.Kinds
+import Language.Granule.Checker.KindsHelpers
 import Language.Granule.Checker.Monad
 import Language.Granule.Checker.Predicates
 import Language.Granule.Checker.SubstitutionContexts
